proper gameticker
